+++ flask_mongoengine/wtf/orm.py
@@ -169,7 +169,7 @@ class ModelConverter(object):
         kwargs = {
             'validators': [],
             'filters': [],
-			'default': field.default or field.document_type_obj,
+            'default': field.default or field.document_type_obj,
         }
         form_class = model_form(field.document_type_obj, field_args={})
         return f.FormField(form_class, **kwargs)